refactor: Move logic to handler
