+++ common/src/main/java/com/wynntils/core/components/Models.java
@@ -32,10 +32,10 @@ import com.wynntils.models.gear.GearTooltipModel;
 import com.wynntils.models.horse.HorseModel;
 import com.wynntils.models.ingredients.IngredientModel;
 import com.wynntils.models.items.ItemModel;
-import com.wynntils.models.label.LabelModel;
 import com.wynntils.models.lootrun.LootrunModel;
 import com.wynntils.models.marker.MarkerModel;
 import com.wynntils.models.mobtotem.MobTotemModel;
+import com.wynntils.models.npc.NpcModel;
 import com.wynntils.models.objectives.ObjectivesModel;
 import com.wynntils.models.particle.ParticleModel;
 import com.wynntils.models.players.FriendsModel;
@@ -86,10 +86,10 @@ public final class Models {
     public static final HorseModel Horse = new HorseModel();
     public static final IngredientModel Ingredient = new IngredientModel();
     public static final ItemModel Item = new ItemModel();
-    public static final LabelModel Label = new LabelModel();
     public static final LootChestModel LootChest = new LootChestModel();
     public static final MarkerModel Marker = new MarkerModel();
     public static final MobTotemModel MobTotem = new MobTotemModel();
+    public static final NpcModel NpcModel = new NpcModel();
     public static final ObjectivesModel Objectives = new ObjectivesModel();
     public static final ParticleModel Particle = new ParticleModel();
     public static final PartyModel Party = new PartyModel();
