+++ common/src/main/java/com/wynntils/handlers/labels/event/LabelIdentifiedEvent.java
@@ -2,9 +2,9 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.label.event;
+package com.wynntils.handlers.labels.event;
 
-import com.wynntils.models.label.type.LabelInfo;
+import com.wynntils.handlers.labels.type.LabelInfo;
 import net.minecraftforge.eventbus.api.Event;
 
 public class LabelIdentifiedEvent extends Event {
