+++ common/src/main/java/com/wynntils/handlers/labels/type/LabelParser.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.label.type;
+package com.wynntils.handlers.labels.type;
 
 import com.wynntils.core.text.StyledText;
 import com.wynntils.utils.mc.type.Location;
