+++ common/src/main/java/com/wynntils/models/profession/label/GatheringNodeLabelInfo.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.label.infos;
+package com.wynntils.models.profession.label;
 
 import com.wynntils.core.text.StyledText;
 import com.wynntils.models.profession.type.MaterialProfile;
