+++ common/src/main/java/com/wynntils/models/profession/label/ProfessionLabelInfo.java
@@ -2,10 +2,10 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.label.infos;
+package com.wynntils.models.profession.label;
 
 import com.wynntils.core.text.StyledText;
-import com.wynntils.models.label.type.LabelInfo;
+import com.wynntils.handlers.labels.type.LabelInfo;
 import com.wynntils.models.profession.type.ProfessionType;
 import com.wynntils.utils.mc.type.Location;
